fix: default limit to max vulnerable paths per vuln, add override option
In CLI-261 an out-of-memory condition occurs when trying to report SCA
findings on a container test. This happens due to an unusually large
number of dependency paths to some vulnerabilities.

Such a large number of paths is unlikely to be useful or helpful.

To mitigate this, vulnerable paths is limited to a maximum of 20 per
vulnerability, which can be raised or lowered using a new
--max-vulnerable-paths option.

If the user has specifically requested all vulnerable paths with
--show-vulnerable-paths=all, this takes precedence and the limit is
ignored.

With these options and additional documentation, OOM issues will be less
likely to occur with default settings, and users interested in vulnerable
path information have more flexibility to mitigate memory issues.